// kilakitu_temp/src/app/model/user.ts
export interface User {
    id?: number;
    type?: string;
    title?: string;
    guests?: number;
    bathrooms?: number;
    beds?: number;
    cost?: number;
    stay?: string;
    units?: string;
    waterbill?: number;
    garbagebill?: number;
    description?: string;
    cancellationpolicy?: string;
    name?: string;
    email?: string;
    phone?: string;
    role?: string;
    dateregistered?: string;
}